requirements.txt
pydantic
python-dotenv
llama-cloud-services
pandas
matplotlib
Flask==2.3.3
Werkzeug==2.3.7
python-dateutil==2.8.2
